implemented randomizing the board
have not implemented game logic yet.
Also I entirely misunderstood how => functions worked and switched back to the old style of functions.
Changed the create 2d array function to allow for 2 parameters, allowing for oddly sized arrays.
Added a draw cell function, which draws a cell based on a value referenced from the currentGrid.
Added "flattenCoordinate", to round input coordinates to the cell cize in anticipation of adding click to change cell value. Also useful for the draw cell function, allowing us to start at a correct coordinate.
Implemented randomize board, allowing one to click and randomize the current board (and subsequently draw it to the screen.)
Added a drawBoard function, which simply draws the board using the drawCell function.